Implementation of Service class to calculate ROE and final amount

+++ src/main/java/com/hackaton/cbft/Constants/Constants.java
@@ -1,10 +1,12 @@
 package com.hackaton.cbft.Constants;
 
+import java.math.BigDecimal;
+
 public class Constants {
 
 	public static final double zeroPFourPercent=0.004;
 	public static final double zeroPThreePercent=0.003;
 	public static final double zeroPTwoPercent=0.002;
 	public static final double zeroPOnePercent=0.001;
-	public static final double DAILY_LIMIT=50000;
+	public static final int DAILY_LIMIT=50000;
 }
